pyproject.toml
[tool.poetry]
name = "gen_tbr_resource"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
webvtt-py = "^0.4.6"
spacy = "^3.0.5"
pymongo = "^3.11.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
gen_review_vocabs = "gen_tbr_resource.entry:file2vocabs"



[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
